feat(ms-movies): add models to movies use case
this commit create enums and models

+++ ms-movies/domain/model/src/main/java/co/com/app/model/movie/gateways/MovieRepository.java
@@ -5,5 +5,6 @@ import reactor.core.publisher.Mono;
 
 public interface MovieRepository {
 
-    Mono<InfoMovie> getMovies(int page, int pageSize);
+    // pageable 1, 5, 10, 20
+    Mono<InfoMovie> getMovies(double page, double pageSize);
 }
